Implement Echidna property testing core setup (Story 024.1)
## Summary
- Install Echidna v2.2.7 locally with functional verification
- Create comprehensive echidna.yaml configuration for property testing
- Implement core property test contracts for TokenLaunch invariants
- Add make echidna target for easy execution
- Document setup process and troubleshooting

## Implementation Details
- Echidna binary installed to ~/.local/bin with pipx crytic-compile
- Basic configuration with Foundry integration and EVM compatibility
- Property test infrastructure including TokenLaunchProperties contract
- Simple validation test demonstrating functional setup
- Comprehensive documentation with patterns and troubleshooting

## Files Added/Modified
- echidna.yaml: Core configuration for property testing
- test/echidna/SimpleTest.sol: Basic functionality validation
- test/echidna/properties/TokenLaunchProperties.sol: Advanced property tests
- Makefile: Added echidna, echidna-coverage, install-echidna targets
- foundry.toml: Updated EVM version for compatibility
- docs/echidna-setup.md: Complete setup and usage documentation

## Validation Evidence
- make echidna successfully runs basic property tests
- Echidna correctly identifies property violations (value >= 1000)
- Integration with Foundry compilation framework confirmed
- All story checklist items completed with functional verification

+++ src/interfaces/IBondingCurveHook.sol
@@ -18,10 +18,12 @@ interface IBondingCurveHook {
      * @return deltaBondingToken Adjustment to bonding token amount (can be positive or negative)
      */
     function buy(
-        address buyer, 
-        uint256 baseBondingToken, 
-        uint256 baseInputToken
-    ) external returns (uint256 fee, int256 deltaBondingToken);
+        address buyer,
+        uint baseBondingToken,
+        uint baseInputToken
+    )
+        external
+        returns (uint fee, int deltaBondingToken);
 
     /**
      * @notice Hook called during remove liquidity operations (sell)
@@ -33,8 +35,10 @@ interface IBondingCurveHook {
      * @return deltaBondingToken Adjustment to required bonding tokens (can be positive or negative)
      */
     function sell(
-        address seller, 
-        uint256 baseBondingToken, 
-        uint256 baseInputToken
-    ) external returns (uint256 fee, int256 deltaBondingToken);
+        address seller,
+        uint baseBondingToken,
+        uint baseInputToken
+    )
+        external
+        returns (uint fee, int deltaBondingToken);
 }
