+++ src/mocks/MockSellHook.sol
@@ -9,68 +9,71 @@ import "../interfaces/IBondingCurveHook.sol";
  * @dev Allows configuration of return values for testing different scenarios
  */
 contract MockSellHook is IBondingCurveHook {
-    uint256 public buyFee;
-    int256 public buyDeltaBondingToken;
-    uint256 public sellFee;
-    int256 public sellDeltaBondingToken;
-    
+    uint public buyFee;
+    int public buyDeltaBondingToken;
+    uint public sellFee;
+    int public sellDeltaBondingToken;
+
     // Call tracking
-    uint256 public buyCallCount;
-    uint256 public sellCallCount;
+    uint public buyCallCount;
+    uint public sellCallCount;
     address public lastBuyer;
     address public lastSeller;
-    uint256 public lastBaseBondingToken;
-    uint256 public lastBaseInputToken;
-    
-    constructor(
-        uint256 _buyFee,
-        int256 _buyDeltaBondingToken,
-        uint256 _sellFee,
-        int256 _sellDeltaBondingToken
-    ) {
+    uint public lastBaseBondingToken;
+    uint public lastBaseInputToken;
+
+    constructor(uint _buyFee, int _buyDeltaBondingToken, uint _sellFee, int _sellDeltaBondingToken) {
         buyFee = _buyFee;
         buyDeltaBondingToken = _buyDeltaBondingToken;
         sellFee = _sellFee;
         sellDeltaBondingToken = _sellDeltaBondingToken;
     }
-    
+
     function buy(
-        address buyer, 
-        uint256 baseBondingToken, 
-        uint256 baseInputToken
-    ) external override returns (uint256 fee, int256 deltaBondingToken) {
+        address buyer,
+        uint baseBondingToken,
+        uint baseInputToken
+    )
+        external
+        override
+        returns (uint fee, int deltaBondingToken)
+    {
         buyCallCount++;
         lastBuyer = buyer;
         lastBaseBondingToken = baseBondingToken;
         lastBaseInputToken = baseInputToken;
-        
+
         return (buyFee, buyDeltaBondingToken);
     }
-    
+
     function sell(
-        address seller, 
-        uint256 baseBondingToken, 
-        uint256 baseInputToken
-    ) external override returns (uint256 fee, int256 deltaBondingToken) {
+        address seller,
+        uint baseBondingToken,
+        uint baseInputToken
+    )
+        external
+        override
+        returns (uint fee, int deltaBondingToken)
+    {
         sellCallCount++;
         lastSeller = seller;
         lastBaseBondingToken = baseBondingToken;
         lastBaseInputToken = baseInputToken;
-        
+
         return (sellFee, sellDeltaBondingToken);
     }
-    
+
     // Configuration functions
-    function setBuyParams(uint256 _fee, int256 _deltaBondingToken) external {
+    function setBuyParams(uint _fee, int _deltaBondingToken) external {
         buyFee = _fee;
         buyDeltaBondingToken = _deltaBondingToken;
     }
-    
-    function setSellParams(uint256 _fee, int256 _deltaBondingToken) external {
+
+    function setSellParams(uint _fee, int _deltaBondingToken) external {
         sellFee = _fee;
         sellDeltaBondingToken = _deltaBondingToken;
     }
-    
+
     function resetCallCounts() external {
         buyCallCount = 0;
         sellCallCount = 0;
